refactor util get_action_type into monitor
